Refactored album paginator and team paginator
